requirements.txt
django==3.1.3
djangorestframework==3.12.1
psycopg2-binary==2.8.6
django-mptt==0.11.0
django-filter==2.3.0
djangorestframework-simplejwt==4.3.0